requirements.txt
Appium-Python-Client~=1.0.2
MonthDelta~=0.9.1
utils~=1.0.1
selenium~=3.141.0
boto3~=1.15.16